Split ConfigObjects into multiple files
