+++ src/main/java/org/p2p/solanaj/rpc/RpcApi.java
@@ -7,9 +7,16 @@ import org.p2p.solanaj.core.Account;
 import org.p2p.solanaj.core.PublicKey;
 import org.p2p.solanaj.core.Transaction;
 import org.p2p.solanaj.rpc.types.*;
-import org.p2p.solanaj.rpc.types.ConfigObjects.*;
+import org.p2p.solanaj.rpc.types.BlockConfig;
+import org.p2p.solanaj.rpc.types.ConfirmedSignFAddr2;
+import org.p2p.solanaj.rpc.types.DataSize;
+import org.p2p.solanaj.rpc.types.Filter;
+import org.p2p.solanaj.rpc.types.Memcmp;
+import org.p2p.solanaj.rpc.types.ProgramAccountConfig;
+import org.p2p.solanaj.rpc.types.RpcEpochConfig;
 import org.p2p.solanaj.rpc.types.RpcResultTypes.ValueLong;
 import org.p2p.solanaj.rpc.types.RpcSendTransactionConfig.Encoding;
+import org.p2p.solanaj.rpc.types.SimulateTransactionConfig;
 import org.p2p.solanaj.rpc.types.TokenResultObjects.*;
 import org.p2p.solanaj.ws.SubscriptionWebSocketClient;
 import org.p2p.solanaj.ws.listeners.NotificationEventListener;
