+++ src/main/java/org/p2p/solanaj/rpc/types/Memcmp.java
@@ -1,0 +1,18 @@
+package org.p2p.solanaj.rpc.types;
+
+import com.squareup.moshi.Json;
+
+public class Memcmp {
+        @Json(name = "offset")
+        private long offset;
+        @Json(name = "bytes")
+        private String bytes;
+
+        public Memcmp() {
+        }
+
+        public Memcmp(long offset, String bytes) {
+            this.offset = offset;
+            this.bytes = bytes;
+        }
+}
